Split code files
